Project25/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
